+++ lib/dartson.dart
@@ -5,6 +5,7 @@ import 'package:logging/logging.dart';
 export 'src/annotations.dart';
 import 'src/annotations.dart';
 import './type_transformer.dart';
+import 'src/reference_mapper.dart';
 
 @MirrorsUsed(metaTargets: const [Property], override: '*')
 import 'dart:mirrors';
@@ -44,22 +45,23 @@ class Dartson<T> {
   /// the type of object. The Class of [object] should have the [Entity]
   /// annotation to work properly when compiling to JavaScript.
   /// Returns the [object].
-  Object map(Object data, Object object, [bool isList = false]) {
+  Object map(Object data, Object object, [bool isList = false, bool beReferenceAware = false]) {
     var reflectee = reflect(object);
+    var mapper = beReferenceAware ? new DecodingReferenceMapper() : null;
 
     if (data is List && isList) {
       ClassMirror itemMirror = reflectee.type;
 
       var list = [];
       data.forEach((item) {
-        list.add(_convertValue(itemMirror, item, "@LIST_ITEM"));
+        list.add(_convertValue(itemMirror, item, "@LIST_ITEM", mapper));
       });
 
       return list;
     } else if (data is List && !isList) {
       throw new IncorrectTypeTransform(object, 'List');
     } else {
-      _fillObject(reflectee, data);
+      _fillObject(reflectee, data, mapper);
     }
 
     return object;
@@ -67,58 +69,81 @@ class Dartson<T> {
 
   /// Transforms an [object] to a serializable map which can be handled
   /// by any encoder like JSON.
-  Object serialize(Object object) {
+  Object serialize(Object object, [EncodingReferenceMapper mapper]) {
     if (object is List) {
-      return _serializeList(object);
+      return _serializeList(object, mapper);
     } else if (object is Map) {
-      return _serializeMap(object);
+      return _serializeMap(object, mapper);
     } else if (object == null) {
       return null;
     } else {
-      return _serializeObject(object);
+      return _serializeObject(object, mapper);
     }
   }
 
-  /// Decodes the [endoded] object (for example a JSON encoded string) using
+  /// Decodes the [encoded] object (for example a JSON encoded string) using
   /// the [_codec] and then uses [map] to map it onto the [object].
-  Object decode(T encoded, Object object, [bool isList = false]) {
-    return map(_codec.decode(encoded), object, isList);
+  /// Reconstruct references and resolved placeholders added by [encodeReferenceAware].
+  Object decodeReferenceAware(T encoded, Object object, [ bool isList = false]) {
+    return decode(encoded, object, isList, true);
+  }
+
+  /// Decodes the [encoded] object (for example a JSON encoded string) using
+  /// the [_codec] and then uses [map] to map it onto the [object].
+  Object decode(T encoded, Object object, [ bool isList = false, bool beReferenceAware =  false]) {
+    return map(_codec.decode(encoded), object, isList, beReferenceAware);
   }
 
   /// Serializes the [decoded] object using [serialize] and then calls the encode
   /// method on the [_codec].
-  T encode(Object decoded) {
-    return _codec.encode(serialize(decoded));
+  /// References to already mapped objects will be serialized by an placeholder (except Lists and Maps)
+  T encodeReferenceAware(Object decoded) {
+    return encode(decoded, beReferenceAware: true);
+  }
+  /// Serializes the [decoded] object using [serialize] and then calls the encode
+  /// method on the [_codec].
+  T encode(Object decoded, {bool beReferenceAware: false}) {
+    return _codec.encode(serialize(decoded, beReferenceAware ? new EncodingReferenceMapper() : null));
   }
 
   /// Creates a new list containing only serializable objects.
-  List _serializeList(List object) {
-    return object.map((i) => serialize(i)).toList();
+  List _serializeList(List object, EncodingReferenceMapper mapper) {
+    return object.map((i) => serialize(i, mapper)).toList();
   }
 
   /// Creates a new map containing only serializable objects.
-  Map _serializeMap(Map object) {
+  Map _serializeMap(Map object, EncodingReferenceMapper mapper) {
     var map = {};
     object.forEach((k, v) {
-      if (v != null) map[k] = serialize(v);
+      if (v != null) map[k] = serialize(v, mapper);
     });
     return map;
   }
 
-  Object _serializeObject(Object object) {
+  Object _serializeObject(Object object, EncodingReferenceMapper mapper) {
+    if (mapper != null && mapper.isSerialized(object)) {
+        return mapper.createSerializablePlaceholder(object);
+    }
     var reflectee = reflect(object);
     var symbolName = _getName(reflectee.type.qualifiedName),
         transformer;
 
     if ((transformer = _transformers[symbolName]) != null) {
-      return transformer.encode(object);
+      var result = transformer.encode(object);
+      if (mapper != null) {
+        mapper.registerSerializableMap(object, result);
+      }
+      return result;
     } else {
       Map result = new Map<String, Object>();
+      if (mapper != null) {
+        mapper.registerSerializableMap(object, result);
+      }
       reflectee.type.declarations.forEach((sym, decl) {
         if (!decl.isPrivate &&
             ((decl is VariableMirror && !decl.isConst && !decl.isStatic) ||
                 (decl is MethodMirror && decl.isGetter))) {
-          _setField(sym, decl, reflectee, result);
+          _setField(sym, decl, reflectee, result, mapper);
         }
       });
 
@@ -131,7 +156,7 @@ class Dartson<T> {
   /// the value to the [result] map. If there's no [Property] annotation
   /// with a different name set it will use the name of [symbol].
   void _setField(Symbol symbol, DeclarationMirror variable,
-      InstanceMirror instMirror, Map<String, Object> result) {
+      InstanceMirror instMirror, Map<String, Object> result, EncodingReferenceMapper mapper) {
     InstanceMirror field = instMirror.getField(symbol);
     Object value = field.reflectee;
     String fieldName = MirrorSystem.getName(symbol);
@@ -148,14 +173,18 @@ class Dartson<T> {
 
     if (value != null && (prop != null ? !prop.ignore : true)) {
       _log.finer("Serializing field: ${fieldName}");
-      result[fieldName] = serialize(value);
+      result[fieldName] = serialize(value, mapper);
     }
   }
 
   /// Puts the data of the [filler] into the object in [objMirror]
   /// Throws [IncorrectTypeTransform] if json data types doesn't match.
-  void _fillObject(InstanceMirror objMirror, Map filler) {
+  void _fillObject(InstanceMirror objMirror, Map filler, DecodingReferenceMapper mapper) {
     ClassMirror classMirror = objMirror.type;
+
+    if (mapper != null) {
+       mapper.registerInstanceIfApplicable(objMirror.reflectee, filler);
+    }
 
     classMirror.declarations.forEach((sym, decl) {
       if (!decl.isPrivate &&
@@ -186,10 +215,12 @@ class Dartson<T> {
             'Try to fill object with: ${fieldName}: ${filler[fieldName]}');
         if (filler[fieldName] != null) {
           objMirror.setField(new Symbol(varName),
-              _convertValue(valueType, filler[fieldName], varName));
+              _convertValue(valueType, filler[fieldName], varName, mapper));
         }
       }
     });
+
+
 
     _log.fine("Filled object completly: ${filler}");
   }
@@ -198,7 +229,14 @@ class Dartson<T> {
   /// returns Deserialized value
   ///  Throws [IncorrectTypeTransform] if json data types doesn't match.
   ///  Throws [NoConstructorError]
-  Object _convertValue(TypeMirror valueType, Object value, String key) {
+  Object _convertValue(TypeMirror valueType, Object value, String key, DecodingReferenceMapper mapper) {
+    if (mapper != null && mapper.isPlaceholder(value)) {
+      var instance = mapper.resolveReferenceForPlaceholder(value);
+      if (instance == null) {
+        _log.severe("unable to resolve reference for placeholder ${value}");
+      }
+      return instance;
+    }
     var symbolName = _getName(valueType.qualifiedName),
         transformer;
 
@@ -212,10 +250,10 @@ class Dartson<T> {
       _log.finer('Handle generic type.');
       // handle generic lists
       if (varMirror.originalDeclaration.qualifiedName == _QN_LIST) {
-        return _convertGenericList(varMirror, value);
+        return _convertGenericList(varMirror, value, mapper);
       } else if (varMirror.originalDeclaration.qualifiedName == _QN_MAP) {
         // handle generic maps
-        return _convertGenericMap(varMirror, value);
+        return _convertGenericMap(varMirror, value, mapper);
       }
     } else if (symbolName == 'dynamic') {
       // dynamic is used in JavaScript runtime
@@ -227,7 +265,7 @@ class Dartson<T> {
       var obj = _initiateClass(valueType);
 
       if (!(value is String) && !(value is num) && !(value is bool)) {
-        _fillObject(obj, value);
+        _fillObject(obj, value, mapper);
       } else {
         throw new IncorrectTypeTransform(value, symbolName, key);
       }
@@ -239,14 +277,14 @@ class Dartson<T> {
   }
 
   /// Converts a list of objects to a list with a Class.
-  List _convertGenericList(ClassMirror listMirror, List fillerList) {
+  List _convertGenericList(ClassMirror listMirror, List fillerList, DecodingReferenceMapper mapper) {
     _log.finer('Converting generic list');
     ClassMirror itemMirror = listMirror.typeArguments[0];
     InstanceMirror resultList = _initiateClass(listMirror);
 
     fillerList.forEach((item) {
       (resultList.reflectee as List)
-          .add(_convertValue(itemMirror, item, "@LIST_ITEM"));
+          .add(_convertValue(itemMirror, item, "@LIST_ITEM", mapper));
     });
 
     _log.finer("Created generic list: ${resultList.reflectee}");
@@ -254,7 +292,7 @@ class Dartson<T> {
   }
 
   /// Converts a generic map.
-  Map _convertGenericMap(ClassMirror mapMirror, Map fillerMap) {
+  Map _convertGenericMap(ClassMirror mapMirror, Map fillerMap, DecodingReferenceMapper mapper) {
     _log.finer('Converting generic map');
 
     ClassMirror itemMirror = mapMirror.typeArguments[1];
@@ -263,8 +301,8 @@ class Dartson<T> {
     Map reflectee = {};
 
     fillerMap.forEach((key, value) {
-      var keyItem = _convertValue(keyMirror, key, "@MAP_KEY");
-      var valueItem = _convertValue(itemMirror, value, "@MAP_VALUE");
+      var keyItem = _convertValue(keyMirror, key, "@MAP_KEY", mapper);
+      var valueItem = _convertValue(itemMirror, value, "@MAP_VALUE", mapper);
       reflectee[keyItem] = valueItem;
 
       _log.finer("Added item ${valueItem} to map key: ${keyItem}");